refactor(notifications): move endpoints to Features/ and slice Contracts
Endpoints/Notifications/* moved to Features/Notifications/<Op>/. The feature-
scoped QueryNotificationsRequest moved to Contracts/Features/Notifications/List/.
Entity, value-object, and channel-payload contracts remain at Contracts root
to avoid migration-snapshot churn (revisit in a follow-up phase).
