+++ modules/Notifications/src/SimpleModule.Notifications/Features/Notifications/UnreadCount/UnreadCountEndpoint.cs
@@ -7,7 +7,7 @@ using SimpleModule.Core.Extensions;
 using SimpleModule.Notifications.Contracts;
 using SimpleModule.Users.Contracts;
 
-namespace SimpleModule.Notifications.Endpoints.Notifications;
+namespace SimpleModule.Notifications.Features.Notifications.UnreadCount;
 
 public class UnreadCountEndpoint : IEndpoint
 {
