+++ modules/Notifications/src/SimpleModule.Notifications/Infrastructure/NotificationService.cs
@@ -1,6 +1,7 @@
 using Microsoft.EntityFrameworkCore;
 using SimpleModule.Core;
 using SimpleModule.Notifications.Contracts;
+using SimpleModule.Notifications.Contracts.Features.Notifications.List;
 using SimpleModule.Users.Contracts;
 
 namespace SimpleModule.Notifications.Infrastructure;
